extract one at a time questions into a separate live component
